% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen.archaeology.R
\docType{package}
\name{allen.archaeology}
\alias{allen.archaeology}
\title{allen.archaeology: A package for applying Allen's interval algebra to Bayesian calibration output}
\description{
The allen.archaeology package provides three categories of functions:
illustrate, plot, and describe.
}
\section{Illustrate function}{

The illustrate function ...
}

\section{Plot functions}{

The plot functions ...
}

\section{Describe function}{

The describe function ...
}

