% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{allen.ggplot2.graphic.full}
\alias{allen.ggplot2.graphic.full}
\title{Make a plot with panels of Nökel lattices.}
\usage{
allen.ggplot2.graphic.full(
  allen.set,
  file.name = NULL,
  pad = 0.2,
  font.size = 11,
  height = 7,
  width = 7
)
}
\arguments{
\item{allen.set}{a dataframe with plot information, such as the one
produced by xxx}

\item{file.name}{optional path to the graphic file output}

\item{pad}{pad the left and right margins to keep labels from disappearing}

\item{font.size}{font size for the labels in the plot}

\item{height}{height in inches of the graphic file output}

\item{width}{width in inches of the graphic file output}
}
\value{
called for its side effects
}
\description{
Plot panels of Nökel lattices to the display and optionally to a file.
}
\author{
Thomas S. Dye
}
