% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{allen.nokel.lattice}
\alias{allen.nokel.lattice}
\title{Plot a Nökel lattice}
\usage{
allen.nokel.lattice(
  allen.set,
  file.name = NULL,
  graph.title = "N\\\\u00F6kel lattice",
  pad = 0.2,
  font.size = 11,
  height = 7,
  width = 7
)
}
\arguments{
\item{allen.set}{a dataframe, such as the one produced by
illustrate.allen.relations()}

\item{file.name}{path to the optional graphics file}

\item{graph.title}{title of the graph}

\item{pad}{pad the left and right margins to keep labels from disappearing}

\item{font.size}{font size for the labels in the plot}

\item{height}{height in inches of the plot saved to file}

\item{width}{width in inches of the plot saved to file}
}
\value{
called for its side effects
}
\description{
Plot a single N\\u00F6kel lattice to the display and optionally to a file.
}
\author{
Thomas S. Dye
}
