% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{allen.proportion.results}
\alias{allen.proportion.results}
\title{Calculate the proportion of each relation in a result vector}
\usage{
allen.proportion.results(result.vector)
}
\arguments{
\item{result.vector}{A result vector}
}
\value{
A named vector with proportions
}
\description{
Divides through by the sum of observations in the result vector.
Assigns the names of the result vector to the return vector.
}
\author{
Thomas S. Dye
}
