% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{allen_composite_relation}
\alias{allen_composite_relation}
\title{Calculate the composite relation of two phases}
\usage{
allen_composite_relation(
  mcmc,
  phases,
  title = c("first", "second"),
  app = "bcal",
  quiet = "partial"
)
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases}{a vector with six column indices representing the start
and end chains of three phases}

\item{title}{a plot title}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}

\item{quiet}{One of "no" to allow messages and warnings,
"partial" (default) to suppress messages and allow warnings, or "yes"
to suppress messages and warnings.}
}
\value{
A dataframe for input to allen_plot_single()
}
\description{
Calculate the composite relation of two phases
}
