% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{allen_plot_multiple}
\alias{allen_plot_multiple}
\title{Make a plot with panels of N\\u00F6kel lattices.}
\usage{
allen_plot_multiple(
  allen_set,
  file_name = NULL,
  pad = 0.2,
  font_size = 11,
  height = 7,
  width = 7
)
}
\arguments{
\item{allen_set}{a dataframe with plot information, such as the one
produced by xxx}

\item{file_name}{optional path to the graphic file output}

\item{pad}{pad the left and right margins to keep labels from disappearing}

\item{font_size}{font size for the labels in the plot}

\item{height}{height in inches of the graphic file output}

\item{width}{width in inches of the graphic file output}
}
\value{
called for its side effects
}
\description{
Plot panels of N\\u00F6kel lattices to the display and optionally to a file.
}
\author{
Thomas S. Dye
}
