% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{allen_relate_phases}
\alias{allen_relate_phases}
\title{Relate phases}
\usage{
allen_relate_phases(mcmc, phases_1, phases_2, app = "bcal")
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases_1}{vector of column indices for the
beginning and end of one or more phases}

\item{phases_2}{vector of column indices for the beginning and end
of one or more phases}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}
}
\value{
A dataframe suitable for plotting with ggraph
}
\description{
Reads MCMC output to create a dataframe suitable
for plotting the Allen relations between two sets of phases
using the six-value logic for intervals with distinct
endpoints.
}
\author{
Thomas S. Dye
}
