% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{allen_relate_stratification_culture}
\alias{allen_relate_stratification_culture}
\title{Dataframe for comparing one or more stratigraphic phases
with a cultural phase.}
\usage{
allen_relate_stratification_culture(
  chronomodel_mcmc,
  stratigraphic_phases,
  cultural_phase
)
}
\arguments{
\item{chronomodel_mcmc}{path to a csv file with Chronomodel MCMC output}

\item{stratigraphic_phases}{vector of column indices, typically for the
beginning and end of one or more stratigraphic phases}

\item{cultural_phase}{vector of column indices for the beginning and end
of a cultural phase}
}
\value{
A dataframe suitable for plotting with ggraph
}
\description{
Reads MCMC output from Chronomodel to create a dataframe suitable
for plotting the Allen relations between stratigraphic phases and
a cultural phase using the six-value logic for phases with distinct
endpoints.
}
\author{
Thomas S. Dye
}
