% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{allen_relation_summary}
\alias{allen_relation_summary}
\title{Summarize the relation of two phases}
\usage{
allen_relation_summary(mcmc, phases, app = "chronomodel")
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases}{a vector with four column indices representing the start
and end of two phases}

\item{app}{one of 'bcal', 'OxCal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}
}
\value{
a list with the following components:
}
\description{
Summarize the relation of two phases
}
