% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{illustrate.allen.relations}
\alias{illustrate.allen.relations}
\title{Prepare data for an illustrative graphic}
\usage{
illustrate.allen.relations(subset = "none")
}
\arguments{
\item{subset}{One of "none" (default),
"concurrent" (two highlight concurrent relations),
or "six" (to highlight relations with distinct endpoints)}
}
\value{
A dataframe for input to ggraph
}
\description{
Create a dataframe that can be used as input for an illustrative plot.
Useful for describing the Allen operators.  Able to illustrate the full
set of Allen relations, concurrent Allen relations, and relations with
distinct endpoints (six-value set).
}
\author{
Thomas S. Dye
}
