% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.R
\name{illustrate}
\alias{illustrate}
\title{Illustrate basic and composite relations}
\usage{
illustrate(relations = "basic", ...)
}
\arguments{
\item{relations}{One of:
\describe{
\item{basic}{show the 13 basic Allen relations (default);}
\item{concurrent}{show concurrent relations;}
\item{six}{show relations with distinct endpoints;}
\item{inherit}{show chronological information about a transmission
    relationship from the receiver's point of view;}
\item{donate}{show chronological information from the giver's point of view during
    transmission;}
\item{stratigraphic}{show the basic stratigraphic relations established by an
    observation of superposition;}
\item{anagenetic}{show the basic relations of artifact change without
    branching;}
\item{cladogenetic}{show the basic relations of artifact change with branching;}
\item{donors}{show the composite relations of two donors to a borrowing
    occurrence;}
\item{inheritors}{show the composite relations of two entities produced by
    innovation from a single source;}
\item{tradition}{show the composite relations of tradition during
    cladogenesis;}
\item{sequence}{show the composite relations of superposition
    in a stratigraphic sequence; or}
\item{anagenesis}{show the composite relations of transmission during
    anagenesis.}}}

\item{...}{Named arguments to be passed on to \code{allen_plot_single()}.}
}
\value{
NULL.  Called for its side effects.
}
\description{
Illustrate basic and composite Allen relations for several chronological
model domains with a N\\u00F6kel lattice. Chronological model domains include
stratigraphy, anagenetic artifact change, and a variety of cladogenetic
processes of artifact change, including what archaeologists label tradition,
innovation, and borrowing. The illustrations show that stratigraphy and
anagenetic artifact change each offer the necessary information to structure
a chronological model deductively. In contrast, cladogenetic processes of
artifact change comprise relatively weak sources of information that offer
little guidance for the chronological model builder.
}
\author{
Thomas S. Dye
}
