% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{allen_plot_single}
\alias{allen_plot_single}
\title{Make a single plot of a Nökel lattice.}
\usage{
allen_plot_single(
  allen_set,
  file_name = NULL,
  pad = 0.2,
  font_size = 11,
  height = 7,
  width = 7,
  title = allen_set$title
)
}
\arguments{
\item{allen_set}{a dataframe with plot information, such as the one
produced by illustrate_allen_relations()}

\item{file_name}{optional path to the graphic file output}

\item{pad}{padding in inches to the margins to keep
labels from disappearing off the edge of the graphic}

\item{font_size}{font size for the labels in the plot}

\item{height}{height in inches of the graphic file output}

\item{width}{width in inches of the graphic file output}

\item{title}{title for the plot, defaults to the title in allen_set.}
}
\value{
called for its side effects
}
\description{
Plots a Nökel lattice to the display and optionally to a file.
}
\author{
Thomas S. Dye
}
