% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{illustrate_allen_relations}
\alias{illustrate_allen_relations}
\title{Data for an illustrative graphic}
\usage{
illustrate_allen_relations(subset = "none")
}
\arguments{
\item{subset}{One of "none" (default),
"concurrent" to highlight concurrent relations,
"six" to highlight relations with distinct endpoints,
"inherit" to highlight chronological information in an inheritance
relationship, or
"donate" to highlight chronological information in a donation relationship.}
}
\value{
A dataframe for input to allen_plot_single()
}
\description{
Create a dataframe that can be used as input for an illustrative plot.
Useful for describing the Allen operators.  Able to illustrate the full
set of Allen relations, concurrent Allen relations, and relations with
distinct endpoints (six-value set).
}
\author{
Thomas S. Dye
}
